Moving types into their own file
