+++ src/structure/tfc/mod.rs
@@ -1,10 +1,12 @@
 pub mod block;
+pub mod datatypes;
 pub mod decimal;
 pub mod dict;
+pub mod file;
 pub mod integer;
 pub mod typed;
-pub mod file;
 
-pub use typed::*;
-pub use block::{SizedDictEntry, SizedDictEntryBuf, OwnedSizedDictEntryBuf};
+pub use block::{OwnedSizedDictEntryBuf, SizedDictEntry, SizedDictEntryBuf};
+pub use datatypes::*;
 pub use file::*;
+pub use typed::*;
